feat: 해결하기 api 연결 80%

+++ app/src/main/java/com/issueissyu/fe/data/remote/dto/response/pin/ProblemSolverJoinResponse.kt
@@ -2,6 +2,6 @@ package com.issueissyu.fe.data.remote.dto.response.pin
 
 data class ProblemSolverJoinResponse (
     val pinId: Long,
-    val problemSolverId: Long,
+    val problemSolverId: Int,
     val problemSolveState: String
 )
